+++ CHANGES.rst
@@ -11,6 +11,8 @@ Unreleased
     the ``Context.invoke()`` method. :issue:`3066` :issue:`3065` :pr:`3068`
 -   Fix conversion of ``Sentinel.UNSET`` happening too early, which caused incorrect
     behavior for multiple parameters using the same name. :issue:`3071` :pr:`3079`
+-   When ``Sentinel.UNSET`` is found during parsing, it will skip calls to
+    ``type_cast_value``. :issue:`3069`
 
 Version 8.3.0
 --------------